correction de bug dans ctrl+[up|down] et modif size_t -> int

+++ libft/srcs/ft_bzero.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void			ft_bzero(void *s, size_t n)
+void			ft_bzero(void *s, int n)
 {
-	size_t		i;
+	int		i;
 	char		*tmp;
 
 	i = 0;
