+++ libft/srcs/ft_strlcat.c
@@ -13,10 +13,10 @@
 #include "libft.h"
 #include <string.h>
 
-size_t			ft_strlcat(char *dst, const char *src, size_t size)
+int			ft_strlcat(char *dst, const char *src, int size)
 {
-	size_t		len_dst;
-	size_t		len_src;
+	int		len_dst;
+	int		len_src;
 
 	len_dst = ft_strlen(dst);
 	len_src = ft_strlen(src);
