+++ libft/srcs/ft_strnew.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-char	*ft_strnew(size_t size)
+char	*ft_strnew(int size)
 {
 	return ((char *)ft_memalloc(size + 1));
 }
